test-requirements.txt
-rrequirements.txt
faker
pytest
pytest-cov
WebTest
hypothesis
docker
gitpython
requests
